requirements.txt
gradio
torch
torchvision
timm
librosa
opencv-python-headless
numpy
Pillow
